+++ CHANGELOG.md
@@ -5,6 +5,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Breaking change: Rename `onClose` to `onAfterClose` in `CartRewardCard`.
+- Breaking change: `mediaQueries` don't exposed all methods.
+  You must use `exposedMethods` prop to list methods who must be exposed.
+
+```js
+mediaQueries(Component, {
+  viewportIsMobile: true,
+  exposedMethods: ['close'],
+})
+```
+
 ## [21.3.1] - 2018-08-02
 
 Fix: